requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
Pillow==9.0.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
sqlparse==0.4.2
